Made the camera persistent **PLEASE READ LAUREN**
I made the camera persistent.  The general definition for persitent in
gamemaker is an object that follows the user through every game maker
room, even when said object is not placed directly in a room.
This allows us to have it in only the start menu room, and it will
follow to every room after that. It also allows us to basically use it
as a storage for variables that we need in the menu's (stockpile, which
levels have been beaten, ect).